Move media span method to line block formatter
