+++ app/src/main/kotlin/org/wordpress/aztec/demo/MainActivity.kt
@@ -100,7 +100,7 @@ class MainActivity : AppCompatActivity(), OnMediaOptionSelectedListener, OnReque
             }
 
             val source = "<img src=\"$mediaPath\">"  // Temporary source value.  Replace with URL after uploaded.
-            aztec.insertMedia(this, BitmapDrawable(resources, bitmap), source)
+            aztec.lineBlockFormatter.insertMedia(this, BitmapDrawable(resources, bitmap), source)
         }
 
         super.onActivityResult(requestCode, resultCode, data)
