+++ aztec/src/main/kotlin/org/wordpress/aztec/formatting/LineBlockFormatter.kt
@@ -1,5 +1,7 @@
 package org.wordpress.aztec.formatting
 
+import android.content.Context
+import android.graphics.drawable.Drawable
 import android.support.v4.content.ContextCompat
 import android.text.Spanned
 import android.text.TextUtils
@@ -7,9 +9,7 @@ import org.wordpress.aztec.AztecText
 import org.wordpress.aztec.R
 import org.wordpress.aztec.TextChangedEvent
 import org.wordpress.aztec.TextFormat
-import org.wordpress.aztec.spans.AztecBlockSpan
-import org.wordpress.aztec.spans.AztecCommentSpan
-import org.wordpress.aztec.spans.AztecHeadingSpan
+import org.wordpress.aztec.spans.*
 import java.util.*
 
 
@@ -289,5 +289,43 @@ class LineBlockFormatter(editor: AztecText) : AztecFormatter(editor) {
         editor.setSelection(commentEndIndex + 1)
     }
 
+    fun insertMedia(context: Context, drawable: Drawable, source: String) {
+        val span = AztecMediaSpan(context, drawable, source)
 
+        //check if we add media into a block element, at the end of the line, but not at the end of last line
+        var applyingOnTheEndOfBlockLine = false
+        editableText.getSpans(selectionStart, selectionEnd, AztecBlockSpan::class.java).forEach {
+            if (editableText.getSpanEnd(it) > selectionEnd && editableText[selectionEnd] == '\n') {
+                applyingOnTheEndOfBlockLine = true
+                return@forEach
+            }
+        }
+
+        val mediaStartIndex = selectionStart + 1
+        val mediaEndIndex = selectionStart + span.getHtml().length + 1
+
+        editor.disableTextChangedListener()
+        editableText.replace(selectionStart, selectionEnd, "\n" + span.getHtml() + if (applyingOnTheEndOfBlockLine) "" else "\n")
+
+        editor.removeBlockStylesFromRange(mediaStartIndex, mediaEndIndex + 1, true)
+        editor.removeHeadingStylesFromRange(mediaStartIndex, mediaEndIndex + 1)
+        editor.removeInlineStylesFromRange(mediaStartIndex, mediaEndIndex + 1)
+
+        editableText.setSpan(
+                span,
+                mediaStartIndex,
+                mediaEndIndex,
+                Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
+        )
+
+        editableText.setSpan(
+                AztecMediaClickableSpan(span),
+                mediaStartIndex,
+                mediaEndIndex,
+                Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
+        )
+
+        editor.setSelection(mediaEndIndex + 1)
+        editor.isMediaAdded = true
+    }
 }
